286-subset_version-should-be-3-digits-for-ddic-and-icsnr-datasets

+++ src/masschange/ingest/executor/datafilereaders/gracefo/primary/ddic.py
@@ -38,7 +38,7 @@ class GraceFODdicDataFileReader(AsciiDataFileReader):
             AsciiDataFileReaderColumn(index=0, name='time', np_type=np.double, unit='s'),
             AsciiDataFileReaderColumn(index=1, name='ddic', np_type=np.double, unit='cycles',
                                       aggregations=['min', 'max']),
-            DerivedAsciiDataFileReaderColumn(name='subset_version', np_type=int, unit=None)
+            DerivedAsciiDataFileReaderColumn(name='subset_version', np_type=np.uint16, unit=None)
         ]
 
     @classmethod
@@ -60,7 +60,7 @@ class GraceFODdicDataFileReader(AsciiDataFileReader):
         """Extract subset version from input file name"""
         filename = os.path.split(filepath)[-1]
         pattern = cls.get_applicable_regex_pattern(filename)
-        return int(re.search(pattern, filename).group('subset_version'))
+        return np.uint16(re.search(pattern, filename).group('subset_version'))
 
     @classmethod
     def load_data_from_file(cls, filepath: str) -> pd.DataFrame:
